refactor: rename grafana adapter to prometheus
