+++ wingfoil/examples/otlp_metrics.rs
@@ -12,11 +12,11 @@
 //!
 //! # Run this example
 //! OTLP_ENDPOINT=http://localhost:4318 \
-//!     cargo run --example otlp_metrics --features otlp,grafana
+//!     cargo run --example otlp_metrics --features otlp,prometheus
 //! ```
 use std::time::Duration;
-use wingfoil::adapters::grafana::PrometheusExporter;
 use wingfoil::adapters::otlp::{OtlpConfig, OtlpPush};
+use wingfoil::adapters::prometheus::PrometheusExporter;
 use wingfoil::*;
 
 fn main() -> anyhow::Result<()> {
