+++ wingfoil/src/adapters/mod.rs
@@ -9,13 +9,13 @@ pub mod etcd;
 #[cfg(feature = "iceoryx2-beta")]
 #[doc(hidden)]
 pub mod iceoryx2;
-#[cfg(feature = "grafana")]
-pub mod grafana;
 pub mod iterator_stream;
 #[cfg(feature = "kdb")]
 pub mod kdb;
 #[cfg(feature = "otlp")]
 pub mod otlp;
+#[cfg(feature = "prometheus")]
+pub mod prometheus;
 #[cfg(feature = "zmq-beta")]
 #[doc(hidden)]
 pub mod zmq;
